requirements.txt
alembic==0.9.9
appnope==0.1.0
backcall==0.1.0
bcrypt==3.1.4
certifi==2018.4.16
cffi==1.11.5
chardet==3.0.4
click==6.7
decorator==4.3.0
Flask==1.0.2
Flask-Bcrypt==0.7.1
Flask-JWT==0.3.2
Flask-Migrate==2.1.1
Flask-SQLAlchemy==2.3.2
gunicorn==19.8.1
idna==2.6
ipython==6.4.0
ipython-genutils==0.2.0
itsdangerous==0.24
jedi==0.12.0
Jinja2==2.10
Mako==1.0.7
MarkupSafe==1.0
parso==0.2.1
pexpect==4.6.0
pickleshare==0.7.4
prompt-toolkit==1.0.15
psycopg2==2.7.4
ptyprocess==0.5.2
pycparser==2.18
Pygments==2.2.0
PyJWT==1.4.2
python-dateutil==2.7.3
python-dotenv==0.8.2
python-editor==1.0.3
requests==2.18.4
simplegeneric==0.8.1
six==1.11.0
SQLAlchemy==1.2.8
traitlets==4.3.2
urllib3==1.22
wcwidth==0.1.7
Werkzeug==0.14.1
